+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes
 
+### Added
+
+* Add helper function for executing futures `test::block_fn()`
+
+
 ### Changed
 
 * Allow to construct `Data` instances to avoid double `Arc` for `Send + Sync` types.